# docs/api/openapi-m01-collection.yaml
openapi: "3.1.0"
info:
  title: Value Betting API — M01 Odds Collection
  version: 1.0.0
  description: |
    Módulo de Coleta de Odds (M01). Endpoints administrativos para gestão
    de ligas, mercados, polling e monitoramento do pipeline de coleta;
    e endpoint Elite para consulta de snapshots históricos de odds.

    ## Arquitetura
    - Admin API: `admin.valuebetting.com` — acesso restrito por IP (ADR-004)
    - Elite API: `api.valuebetting.com` — acesso via API Key com rate limiting

    ## Referências
    - Requisitos: `docs/requirements/01_odds_collection.md`
    - ADR-004 (API Stack), ADR-007 (Auth), ADR-010 (Estrutura)
  contact:
    name: Value Betting Team
  license:
    name: Proprietary
servers:
  - url: https://admin.valuebetting.com
    description: Admin API (IP allowlist + JWT)
  - url: https://api.valuebetting.com
    description: User/Elite API
...
paths:
...
  /api/v1/admin/collection/leagues:
    get:
      operationId: listLeagues
      summary: Listar ligas habilitadas para coleta
      tags: [Collection Leagues]
      security:
        - AdminBearerAuth: []
      parameters:
        - name: sport
          in: query
          schema:
            $ref: "#/components/schemas/SportEnum"
        - name: active
          in: query
          schema:
            type: boolean
        - name: tierRequired
          in: query
          schema:
            $ref: "#/components/schemas/TierEnum"
      responses:
        "200":
          description: Lista de ligas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      operationId: createLeague
      summary: Habilitar nova liga para coleta
      tags: [Collection Leagues]
      security:
        - AdminBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeagueCreateRequest"
      responses:
        "201":
          description: Liga habilitada com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/admin/collection/leagues/{leagueId}:
    patch:
      operationId: updateLeague
      summary: Atualizar configuração de liga
      tags: [Collection Leagues]
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/LeagueIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaguePatchRequest"
      responses:
        "200":
          description: Liga atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: disableLeague
      summary: Desabilitar liga (soft-delete — active=false)
      description: |
        Desabilita a liga para coleta. Alertas pendentes dessa liga são
        cancelados. A liga permanece no banco para auditoria.
      tags: [Collection Leagues]
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/LeagueIdPath"
      responses:
        "204":
          description: Liga desabilitada
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---- Market Types ----
  /api/v1/admin/collection/market-types:
    get:
      operationId: listMarketTypes
      summary: Listar tipos de mercado configurados
      tags: [Collection Market Types]
      security:
        - AdminBearerAuth: []
      responses:
        "200":
          description: Lista de tipos de mercado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketTypeListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/collection/market-types/{marketTypeId}:
    patch:
      operationId: updateMarketType
      summary: Habilitar/desabilitar tipo de mercado
      tags: [Collection Market Types]
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/MarketTypeIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketTypePatchRequest"
      responses:
        "200":
          description: Tipo de mercado atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketTypeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---- Polling Config ----
  /api/v1/admin/collection/polling/config:
    get:
      operationId: getPollingConfig
      summary: Obter configuração global de polling
      tags: [Collection Polling]
      security:
        - AdminBearerAuth: []
      responses:
        "200":
          description: Configuração atual de polling
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollingConfigResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    patch:
      operationId: updatePollingConfig
      summary: Atualizar intervalos globais de polling
      tags: [Collection Polling]
      security:
        - AdminBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PollingConfigPatchRequest"
      responses:
        "200":
          description: Configuração atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollingConfigResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ---- Polling Overrides ----
  /api/v1/admin/collection/polling/overrides:
    get:
      operationId: listPollingOverrides
      summary: Listar overrides de polling por par liga/bookmaker
      description: |
        Retorna todos os overrides de frequência de polling ativos.
        Cada override define um intervalo customizado para um par
        (liga, bookmaker) conforme RF-COL-A.
      tags: [Collection Polling]
      security:
        - AdminBearerAuth: []
      responses:
        "200":
          description: Lista de overrides
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollingOverrideListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      operationId: createPollingOverride
      summary: Criar override de polling para par liga/bookmaker
      tags: [Collection Polling]
      security:
        - AdminBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PollingOverrideCreateRequest"
      responses:
        "201":
          description: Override criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollingOverrideResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/admin/collection/polling/overrides/{leagueId}/{bookmakerId}:
    delete:
      operationId: deletePollingOverride
      summary: Remover override de polling
      tags: [Collection Polling]
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/LeagueIdPath"
        - $ref: "#/components/parameters/BookmakerIdPath"
      responses:
        "204":
          description: Override removido
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---- Health & Quality ----
...
  /api/v1/admin/collection/quality-alerts:
    get:
      operationId: listQualityAlerts
      summary: Listar alertas de qualidade de dados
      description: |
        Retorna alertas de anomalia detectados pelo pipeline (RF-COL-010):
        cobertura Pinnacle baixa, alta volatilidade, bookmaker não confiável.
      tags: [Collection Health]
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: severity
          in: query
          schema:
            type: string
            enum: [info, warning, critical]
        - name: alertType
          in: query
          schema:
            type: string
            enum:
              - pinnacle_coverage_low
              - high_volatility
              - bookmaker_unreliable
              - odds_anomaly
        - $ref: "#/components/parameters/FromParam"
        - $ref: "#/components/parameters/ToParam"
      responses:
        "200":
          description: Lista paginada de alertas de qualidade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QualityAlertListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/collection/bookmakers/reliability:
    get:
      operationId: listBookmakerReliability
      summary: Status de confiabilidade dos bookmakers
      description: |
        Retorna métricas de confiabilidade por bookmaker. Bookmakers com
        cobertura < 10% por 3 coletas consecutivas são marcados como
        unreliable (RF-COL-010).
      tags: [Collection Health]
      security:
        - AdminBearerAuth: []
      responses:
        "200":
          description: Lista de bookmakers com métricas de confiabilidade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmakerReliabilityListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ---- Quota ----
  /api/v1/admin/collection/quota:
    get:
      operationId: getQuotaUsage
      summary: Consumo atual de quota da API
      description: |
        Retorna contador de requisições do mês atual para The Odds API.
        Alertas internos em 80% (warning) e 95% (critical) conforme RF-COL-009.
      tags: [Collection Quota]
      security:
        - AdminBearerAuth: []
      responses:
        "200":
          description: Consumo de quota
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotaUsageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/collection/quota/projection:
    get:
      operationId: getQuotaProjection
      summary: Projeção de consumo de quota
      description: |
        Projeta consumo de quota até o final do mês com base na taxa
        atual de requisições. Inclui recomendação de ação se projeção
        exceder limite.
      tags: [Collection Quota]
      security:
        - AdminBearerAuth: []
      responses:
        "200":
          description: Projeção de quota
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotaProjectionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ---- Degradation ----
  /api/v1/admin/collection/degradation:
    get:
      operationId: getDegradationStatus
      summary: Status de modo degradado
      tags: [Collection Degradation]
      security:
        - AdminBearerAuth: []
      responses:
        "200":
          description: Status de degradação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DegradationStatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/collection/degradation/activate:
    post:
      operationId: activateDegradedMode
      summary: Ativar modo degradado manualmente
      description: |
        Ativa modo degradado do pipeline. Novos alertas de EV são
        suspensos (sem dados = sem alertas falsos). Requer motivo
        para audit trail (RF-COL-008).
      tags: [Collection Degradation]
      security:
        - AdminBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DegradationActivateRequest"
      responses:
        "200":
          description: Modo degradado ativado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DegradationStatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/collection/degradation/deactivate:
    post:
      operationId: deactivateDegradedMode
      summary: Desativar modo degradado
      description: |
        Retoma operação normal do pipeline. Coleta e alertas de EV
        são reativados.
      tags: [Collection Degradation]
      security:
        - AdminBearerAuth: []
      responses:
        "200":
          description: Modo normal restaurado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DegradationStatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ---- Elite Snapshots ----
...
components:
...
  securitySchemes:
    AdminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT RS256 para admin (ADR-007). Claims: sub (admin_id),
        role (ops|super_admin), exp (TTL 4h). Requer MFA TOTP prévio.
        Validado pelo API Gateway JWT Authorizer + FastAPI Depends.

    EliteApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API Key Elite no formato vb_elite_<32chars> (ADR-007).
        Rate limit: 1000 req/hora via API Gateway Usage Plan.
        Hash SHA-256 armazenado. Exibida em plain-text apenas na criação.

  # ---- Reusable Parameters ----
...
  responses:
...
    Unauthorized:
      description: Autenticação necessária ou token/API Key inválido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: UNAUTHORIZED
            message: "Token inválido ou expirado"
            details: null
            traceId: "abc-123-def-456"
    Forbidden:
      description: Sem permissão para este recurso
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: FORBIDDEN
            message: "Permissão insuficiente para este recurso"
            details: null
            traceId: "abc-123-def-456"
...
  schemas:
...
    SportEnum:
      type: string
      enum: [football, tennis, basketball, baseball]
    TierEnum:
      type: string
      enum: [free, pro, elite]
...
    LeagueCreateRequest:
      type: object
      required: [leagueId, name, sport, country]
      properties:
        leagueId:
          type: string
          pattern: "^[a-z0-9_]+$"
          example: premier_league
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: Premier League
        sport:
          $ref: "#/components/schemas/SportEnum"
        country:
          type: string
          minLength: 2
          maxLength: 2
          description: ISO 3166-1 alpha-2
          example: GB
        tierRequired:
          $ref: "#/components/schemas/TierEnum"
          default: pro
        minMarketsPerRound:
          type: integer
          minimum: 1
          maximum: 50
          default: 5
...
    LeagueResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/League"
    LeagueListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/League"
    MarketTypeResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/MarketType"
...